requirements.txt
aws-lambda-powertools
python-dotenv
aws-xray-sdk
